+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.0-beta] - 2018-03-19
+
+### Added
+- Support for Button group field type.
+- Support for Accordion field type.
+- Support for `Return format` setting for User field type.
+
+### Changed
+- Enhanced documentation and corrected a few typos.
+
 ## [1.5.5] - 2018-03-16
 
 ### Fixed